Print list

+++ 0x17-doubly_linked_lists/0-print_dlistint.c
@@ -10,7 +10,7 @@
 
 size_t print_dlistint(const dlistint_t *h)
 {
-	size_t nodes = 0;
+	int nodes = 0;
 
 	if (h == NULL)
 		return (nodes);
